ARM: dts: imx6ull-colibri: reduce v_batt current in power off
Reduce the current drawn from VCC_BATT when the main power on the 3V3 pins
to the module are switched off.

This switches off SoC internal pull resistors which are provided on the module
for TAMPER7 and TAMPER9 SoC pin and switches on a pull down instead of a pullup
for the USBC_DET module pin (TAMPER2).

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>
(cherry picked from commit 9fd59b6eebd12106b5b53fc56466bcd89b9a7ee9)

Conflicts:
	arch/arm/boot/dts/imx6ull-colibri.dtsi

Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>